Remove exhibitors

+++ administrator/components/com_projects/views/exhibitors/view.html.php
@@ -44,10 +44,10 @@ class ProjectsViewExhibitors extends HtmlView
         {
             JToolbarHelper::editList('exhibitor.edit');
         }
-        /*if (ProjectsHelper::canDo('core.general'))
+        if (ProjectsHelper::canDo('projects.access.exhibitors.full'))
         {
             JToolbarHelper::deleteList('COM_PROJECT_QUEST_REMOVE_EXHIBITOR', 'exhibitors.delete');
-        }*/
+        }
 		if (ProjectsHelper::canDo('core.admin'))
 		{
 			JToolBarHelper::preferences('com_projects');
